+++ src/weewx/reportengine.py
@@ -27,7 +27,7 @@ import weeutil.weeutil
 import weewx.defaults
 import weewx.manager
 import weewx.units
-from weeutil.weeutil import to_bool, to_int
+from weeutil.weeutil import getFileName, to_bool, to_int, dict_search
 
 log = logging.getLogger(__name__)
 
@@ -194,7 +194,7 @@ class StdReportEngine(threading.Thread):
                 timing_line = skin_dict.get('report_timing')
                 if timing_line:
                     # Get a ReportTiming object.
-                    timing = ReportTiming(timing_line)
+                    timing = ReportTiming(timing_line, skin_dict, self.record['dateTime'])
                     if timing.is_valid:
                         # Get timestamp and interval, so we can check if the
                         # report timing is triggered.
@@ -668,6 +668,10 @@ class ReportTiming:
         @weekly   : Run once a week,  ie "0 0 * * 0"
         @daily    : Run once a day,   ie "0 0 * * *"
         @hourly   : Run once an hour, ie "0 * * * *"
+    - if the timing is given with @createIfMissing the report generation is
+      allowed to occur if files that the report would generate doesn't exist
+      or if the extension has been updated and the modified file times in the
+      skin directory are newer even if the report_timing would normally fail
 
     Useful ReportTiming class attributes:
 
@@ -679,7 +683,7 @@ class ReportTiming:
                       replaced with numeric equivalents.
     """
 
-    def __init__(self, raw_line):
+    def __init__(self, raw_line, skin_dict, dateTime):
         """Initialises a ReportTiming object.
 
         Processes raw line to produce 5 field line suitable for further
@@ -691,6 +695,10 @@ class ReportTiming:
         # initialise some properties
         self.is_valid = None
         self.validation_error = None
+        self.create_if_missing = False
+        self.skin_dict = skin_dict
+        self.dateTime = dateTime
+
         # To simplify error reporting keep a copy of the raw line passed to us
         # as a string. The raw line could be a list if it included any commas.
         # Assume a string but catch the error if it is a list and join the list
@@ -707,6 +715,11 @@ class ReportTiming:
                 self.validation_error = "Unsupported character '%s' in '%s'." % (unsupported_char,
                                                                                  self.raw_line)
                 return
+
+        if "@createIfMissing" in line_str:
+            self.create_if_missing = True
+            self.raw_line = line_str = line_str.replace(",@createIfMissing", "")
+
         # Six special time definition 'nicknames' are supported which replace
         # the line elements with pre-determined values. These nicknames start
         # with the @ character. Check for any of these nicknames and substitute
@@ -874,6 +887,38 @@ class ReportTiming:
                 checked for triggering. May be omitted in which case only
                 ts_hi is checked.
         """
+
+        if self.is_valid and self.create_if_missing:
+            # check for missing files
+
+            skin_dir = self.skin_dict["SKIN_ROOT"]
+            skin_dir = os.path.join(skin_dir, self.skin_dict["skin"])
+            html_dest_dir = self.skin_dict["HTML_ROOT"]
+
+            if os.path.exists(skin_dir):
+                if os.path.exists(html_dest_dir):
+                    templates = dict_search(self.skin_dict.get("CheetahGenerator", None), "template")
+                    for template in templates:
+                        if not template.endswith(".tmpl"):
+                            continue
+
+                        template_filename = os.path.join(skin_dir, template)
+                        if os.path.exists(template_filename):
+                            output_filename = os.path.join(html_dest_dir, getFileName(template, self.dateTime))
+                            if not os.path.exists(output_filename):
+                                log.debug(f"{output_filename} should exist but doesn't, allowing report generation")
+                                return True
+
+                            template_mtime = os.path.getmtime(template_filename)
+                            output_mtime = os.path.getmtime(output_filename)
+
+                            if template_mtime > output_mtime:
+                                log.debug(f"{output_filename} exists but is older than {template_filename}, allowing report generation")
+                                return True
+
+                else:
+                    log.debug(f"{html_dest_dir} should exist but doesn't, allowing report generation")
+                    return True
 
         if self.is_valid and ts_hi is not None:
             # setup ts range to iterate over