+++ docs_src/changes.md
@@ -5,6 +5,10 @@ WeeWX change history
 
 Initialize the widgets in the Seasons skin as soon as the DOM is ready.
 
+Example `xstats.py` had nonsensical time context values. Fixed in 
+[PR #1072](https://github.com/weewx/weewx/pull/1072). Thanks to user 
+evilbunny2008. 
+
 
 ### 5.3.1 03/03/2026