+++ docs_src/changes.md
@@ -9,6 +9,12 @@ Example `xstats.py` had nonsensical time context values. Fixed in
 [PR #1072](https://github.com/weewx/weewx/pull/1072). Thanks to user 
 evilbunny2008. 
 
+Added @createIfMissing option for report_timing handling so that report
+generation will be allowed if there are missing report files or if the
+modified times of templates are newer than generated files.
+Added in [PR #1076](https://github.com/weewx/weewx/pull/1076). Thanks
+to user evilbunny2008.
+
 
 ### 5.3.1 03/03/2026