+++ docs_src/changes.md
@@ -13,6 +13,8 @@ Prevent `AttributeError` when the `weectl` command is run without a subcommand.
 
 Guard against an empty string being passed to the copy generator.
 
+Use better measure of system uptime.
+
 
 ### 5.3.1 03/03/2026