+++ docs_src/changes.md
@@ -12,7 +12,7 @@ Example `xstats.py` had nonsensical time context values. Fixed in
 [PR #1072](https://github.com/weewx/weewx/pull/1072). Thanks to user 
 evilbunny2008. 
 
-Cheetahgenerator seasons timespans `$season`, `$seasonsyear`, seasons
+Added Cheetah generator seasons timespans `$season`, `$seasonsyear`, and seasons
 iterator `.seasons`. [PR #1095](https://github.com/weewx/weewx/pull/1095).
 Thanks to user roe-dl.